+++ packages/parser/src/vue/template.ts
@@ -5,7 +5,8 @@ import {
   type JSExpression,
   type NodeDirective,
   type BlockSlot,
-  type JSFunction
+  type JSFunction,
+  type PlatformType
 } from '@vtj/core';
 import { compileTemplate } from '@vue/compiler-sfc';
 import {
@@ -21,7 +22,12 @@ import {
 } from '@vue/compiler-core';
 import { uid } from '@vtj/base';
 import { isJSExpression, isNodeSchema } from '../shared';
-import { getJSExpression, getJSFunction, formatTagName } from './utils';
+import {
+  getJSExpression,
+  getJSFunction,
+  formatTagName,
+  styleToJson
+} from './utils';
 import type { CSSRules } from './style';
 import { htmlToNodes } from './html';
 
@@ -29,8 +35,10 @@ let __slots: BlockSlot[] = [];
 let __context: Record<string, Set<string>> = {};
 let __handlers: Record<string, JSFunction> = {};
 let __styles: CSSRules = {};
+let __platform: PlatformType = 'web';
 
 export interface ParseTemplateOptions {
+  platform: PlatformType;
   handlers?: Record<string, JSFunction>;
   styles?: CSSRules;
 }
@@ -45,7 +53,7 @@ export function parseTemplate(
   __context = {};
   __handlers = options?.handlers || {};
   __styles = options?.styles || {};
-
+  __platform = options?.platform || 'web';
   const result = compileTemplate({
     id,
     filename: name,
@@ -78,17 +86,6 @@ function pickSlot(node: NodeSchema) {
       params
     });
   }
-}
-
-function styleToJson(style: string) {
-  const cleaned = style.replace(/\s+/g, ' ');
-  return cleaned.split(';').reduce((acc: Record<string, string>, current) => {
-    const [property, value] = current.split(':').map((item) => item.trim());
-    if (property && value) {
-      acc[property] = value;
-    }
-    return acc;
-  }, {});
 }
 
 function getProps(nodes: Array<AttributeNode | DirectiveNode>) {
@@ -313,7 +310,7 @@ function createNodeSchema(
   scope?: IfNode | ForNode
 ) {
   const dsl: NodeSchema = {
-    name: formatTagName(node.tag),
+    name: formatTagName(node.tag, __platform),
     props: getProps(node.props),
     events: getEvents(node.props, __handlers),
     directives: getDirectives(scope || node)
@@ -386,6 +383,11 @@ function transformNode(
     return transformCompoundExpression(
       node.children as CompoundExpressionNode[]
     );
+  }
+
+  // 注释，忽略
+  if (node.type === NodeTypes.COMMENT) {
+    return null;
   }
 
   console.warn('未处理', node.type);