+++ packages/renderer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.8](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.12.7...@vtj/renderer@0.12.8) (2025-04-29)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
 ## [0.12.7](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.12.6...@vtj/renderer@0.12.7) (2025-04-27)
 
 **Note:** Version bump only for package @vtj/renderer